refactor(custom-views): minor improvements
