+++ models/custom-view/src/custom-view/constants.ts
@@ -1,0 +1,10 @@
+export const defaultLocators = [
+  'products.product_details.general',
+  'products.product_variant_details.general',
+  'categories.category_details.external_search',
+  'customers.customer_addresses.edit_address',
+  'orders.order_details.custom_field',
+  'standalone_prices.standalone_price_details',
+  'discounts.product_discount_details.custom_fields',
+  'operations.import_log_details',
+];
